Delete command confirmation

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -1,5 +1,4 @@
 package seedu.address.logic.commands;
-
 import static java.util.Objects.requireNonNull;
 
 import java.util.List;
@@ -11,24 +10,20 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.ui.MainWindow;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
  */
 public class DeleteCommand extends Command {
-
     public static final String COMMAND_WORD = "delete";
-
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Deletes the person identified by the index number used in the displayed person list.\n"
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + COMMAND_WORD + " 1";
-
     public static final String MESSAGE_DELETE_PERSON_SUCCESS = "Deleted Person: %1$s";
-
     private final Index targetIndex;
     private final Name targetName;
-
     /**
      * Creates a DeleteCommand to delete the person at the specified {@code Index}.
      *
@@ -38,7 +33,6 @@ public class DeleteCommand extends Command {
         this.targetIndex = targetIndex;
         this.targetName = null;
     }
-
     /**
      * Creates a DeleteCommand to delete the person with the specified {@code Name}.
      *
@@ -48,12 +42,10 @@ public class DeleteCommand extends Command {
         this.targetIndex = null;
         this.targetName = targetName;
     }
-
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
         List<Person> lastShownList = model.getFilteredPersonList();
-
         Person personToDelete;
         if (targetIndex != null) {
             if (targetIndex.getZeroBased() >= lastShownList.size()) {
@@ -70,29 +62,28 @@ public class DeleteCommand extends Command {
             personToDelete = personOptional.get();
         }
 
+        boolean isConfirmed = MainWindow.showConfirmationDialog("Are you sure you want to delete "
+                + personToDelete.getName() + "?");
+        if (!isConfirmed) {
+            return new CommandResult("Deletion cancelled.");
+        }
+
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, personToDelete));
     }
-
-
-
     @Override
     public boolean equals(Object other) {
         if (other == this) {
             return true;
         }
-
         // instanceof handles nulls
         if (!(other instanceof DeleteCommand)) {
             return false;
         }
-
         DeleteCommand otherDeleteCommand = (DeleteCommand) other;
         return (targetIndex != null && targetIndex.equals(otherDeleteCommand.targetIndex))
                 || (targetName != null && targetName.equals(otherDeleteCommand.targetName));
     }
-
-
     @Override
     public String toString() {
         if (targetIndex != null) {
